+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * kubernetes: add support for managing clusters
 * kubernetes: add support for kkp project and cluster id and vlans
 * kubernetes: add support for managing nodepools
+* kubernetes: add support for managing oidc fields
+* kubernetes: add support for managing servicevm fields
 
 <!--
 Please add your changelog entry under this comment in the correct category (Security, Fixed, Added, Changed, Deprecated, Removed - in this order).